Remove message store (unused)
