+++ src/util/message_record.rs
@@ -1,0 +1,1 @@
+pub use storage_handler::MessageRecord;
