--- src/util/message_store.rs
@@ -1,3 +1,0 @@
-#![allow(unused_imports)]
-
-pub use storage_handler::{MessageRecord, MessageStore};
